+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix tests about signal handling. Code was already correct, but tests were not.
+
 ### Changed, breaking
 
 ### Changed
 
 ### New
+
+- The sleepit helper now performs signal handling, to enable better tests.
 
 ## [v0.6.0] - [2021-01-17]